move `FuncArgs` to `expr::function::aggregate`
