+++ src/expr/src/function/aggregate/mod.rs
@@ -17,3 +17,6 @@ pub use kind::*;
 
 mod call;
 pub use call::*;
+
+mod args;
+pub use args::*;
