+++ src/expr/src/function/mod.rs
@@ -13,5 +13,4 @@
 // limitations under the License.
 
 pub mod aggregate;
-pub mod args;
 pub mod window;
