+++ src/stream/benches/hash_agg.rs
@@ -18,8 +18,7 @@ use futures::StreamExt;
 use risingwave_common::catalog::{Field, Schema};
 use risingwave_common::types::DataType;
 use risingwave_expr::expr::*;
-use risingwave_expr::function::aggregate::{AggCall, AggKind};
-use risingwave_expr::function::args::FuncArgs;
+use risingwave_expr::function::aggregate::{AggCall, AggKind, FuncArgs};
 use risingwave_storage::memory::MemoryStateStore;
 use risingwave_storage::StateStore;
 use risingwave_stream::executor::test_utils::agg_executor::new_boxed_hash_agg_executor;
