+++ src/stream/src/executor/hash_agg.rs
@@ -738,8 +738,7 @@ pub mod tests {
     use risingwave_common::array::StreamChunk;
     use risingwave_common::catalog::{Field, Schema};
     use risingwave_common::types::DataType;
-    use risingwave_expr::function::aggregate::{AggCall, AggKind};
-    use risingwave_expr::function::args::FuncArgs;
+    use risingwave_expr::function::aggregate::{AggCall, AggKind, FuncArgs};
     use risingwave_storage::memory::MemoryStateStore;
     use risingwave_storage::StateStore;
