+++ src/stream/src/executor/integration_tests.rs
@@ -22,8 +22,7 @@ use risingwave_common::array::*;
 use risingwave_common::catalog::{Field, Schema};
 use risingwave_common::types::*;
 use risingwave_expr::expr::*;
-use risingwave_expr::function::aggregate::{AggCall, AggKind};
-use risingwave_expr::function::args::FuncArgs;
+use risingwave_expr::function::aggregate::{AggCall, AggKind, FuncArgs};
 use risingwave_storage::memory::MemoryStateStore;
 
 use super::exchange::permit::channel_for_test;
